code organization and docstrings
