+++ src/Hamiltonians/ExtendedHubbardReal1D.jl
@@ -7,10 +7,13 @@ Implements the extended Hubbard model on a one-dimensional chain in real space.
 \\hat{H} = -t \\sum_{\\langle i,j\\rangle} a_i^† a_j + \\frac{u}{2}\\sum_i n_i (n_i-1) + v \\sum_{\\langle i,j\\rangle} n_i n_j
 ```
 
-# Parameters
-- `u`: on-site interaction parameter
-- `v`: the next-neighbor interaction
-- `t`: the hopping strength
+# Arguments
+
+* `address`: the starting address.
+* `u`: on-site interaction parameter
+* `v`: the next-neighbor interaction
+* `t`: the hopping strength
+
 """
 struct ExtendedHubbardReal1D{TT,A<:AbstractFockAddress,U,V,T} <: AbstractHamiltonian{TT}
     add::A
@@ -38,18 +41,8 @@ Base.getproperty(h::ExtendedHubbardReal1D{<:Any,<:Any,U}, ::Val{:u}) where U = U
 Base.getproperty(h::ExtendedHubbardReal1D{<:Any,<:Any,<:Any,V}, ::Val{:v}) where V = V
 Base.getproperty(h::ExtendedHubbardReal1D{<:Any,<:Any,<:Any,<:Any,T}, ::Val{:t}) where T = T
 
-function diagME(h::ExtendedHubbardReal1D, b::BoseFS)
-    ebhinteraction, bhinteraction = extended_bose_hubbard_interaction(b)
-    return h.u * bhinteraction / 2 + h.v * ebhinteraction
-end
-
 function numOfHops(::ExtendedHubbardReal1D, address::BoseFS)
-    return numberlinkedsites(address)
-end
-
-function hop(h::ExtendedHubbardReal1D, add::BoseFS, chosen)
-    naddress, onproduct = hopnextneighbour(add, chosen)
-    return naddress, - h.t * sqrt(onproduct)
+    return 2 * numberoccupiedsites(address)
 end
 
 """
@@ -90,4 +83,14 @@ function extended_bose_hubbard_interaction(b::BoseFS{<:Any,M}) where M
     end
     ebhmmatrixelementint+= bosonnumber2 * firstbosonnumber  #periodic bondary condition
     return ebhmmatrixelementint , bhmmatrixelementint
+end
+
+function diagME(h::ExtendedHubbardReal1D, b::BoseFS)
+    ebhinteraction, bhinteraction = extended_bose_hubbard_interaction(b)
+    return h.u * bhinteraction / 2 + h.v * ebhinteraction
+end
+
+function hop(h::ExtendedHubbardReal1D, add::BoseFS, chosen)
+    naddress, onproduct = hopnextneighbour(add, chosen)
+    return naddress, - h.t * sqrt(onproduct)
 end
