+++ src/Hamiltonians/abstract.jl
@@ -1,61 +1,49 @@
-# First we have some generic types and methods for any linear operator
-# that could be used for FCIQMC
-
+###
+### This file contains abstract types, interfaces and traits.
+###
 """
     AbstractHamiltonian{T}
-Supertype that provides an interface for linear operators over a linear space with scalar
-type `T` that are suitable for FCIQMC. Indexing is done with addresses (typically not integers)
-from an address space that may be large (and will not need to be completely generated).
 
-`AbstractHamiltonian` instances operate on vectors of type [`AbstractDVec`](@ref)
-from the module `DictVectors` and work well with addresses of type [`AbstractFockAddress`](@ref)
-from the module `BitStringAddresses`. The type works well with the external package `KrylovKit.jl`.
+Supertype that provides an interface for linear operators over a linear space with scalar
+type `T` that are suitable for FCIQMC. Indexing is done with addresses (typically not
+integers) from an address space that may be large (and will not need to be completely
+generated).
+
+`AbstractHamiltonian` instances operate on vectors of type [`AbstractDVec`](@ref) from the
+module `DictVectors` and work well with addresses of type [`AbstractFockAddress`](@ref) from
+the module `BitStringAddresses`. The type works well with the external package
+[KrylovKit.jl](https://github.com/Jutho/KrylovKit.jl).
+
+# Methods
 
 Provides:
+
 * [`hops`](@ref): iterator over reachable off-diagonal matrix elements
 * [`generateRandHop`](@ref): function to generate random off-diagonal matrix element
-* `hamiltonian[address1, address2]`: indexing with `getindex()` - mostly for testing purposes
-* `*(LO, v)` deterministic matrix-vector multiply (`== LO(v)`)
-* `mul!(w, LO, v)` mutating matrix-vector multiply
-* [`dot(x, LO, v)`](@ref) compute `x⋅(LO*v)` minimizing allocations
+* [`dimension`](@ref): get the dimension of the address space.
+* `H[address1, address2]`: indexing with `getindex()` - mostly for testing purposes
+* `*(H, v)`: deterministic matrix-vector multiply.
+* `H(v)`: equivalent to `LO * v`.
+* `mul!(w, H, v)`: mutating matrix-vector multiply.
+* [`dot(x, H, v)`](@ref): compute `x⋅(LO*v)` minimizing allocations.
 
 Methods that need to be implemented:
+
 * [`numOfHops(::AbstractHamiltonian, address)`](@ref)
 * [`hop(::AbstractHamiltonian, address, chosen::Integer)`](@ref)
 * [`diagME(::AbstractHamiltonian, address)`](@ref)
 * [`starting_address(::AbstractHamiltonian)`](@ref)
-Optional:
+
+Optional methods to implement:
+
 * [`Hamiltonians.LOStructure(::Type{typeof(lo)})`](@ref)
 * [`dimension(::Type{T}, ::AbstractHamiltonian)`](@ref)
 * [`hops(::AbstractHamiltonian, ::AbstractFockAddress)`](@ref)
+* [`momentum(::AbstractHamiltonian)`](@ref)
 """
 abstract type AbstractHamiltonian{T} end
 
 Base.eltype(::AbstractHamiltonian{T}) where {T} = T
-
-function Base.:*(h::AbstractHamiltonian{E}, v::AbstractDVec{K,V}) where {E, K, V}
-    T = promote_type(E,V) # allow for type promotion
-    w = empty(v, T) # allocate new vector; non-mutating version
-    for (key,val) in pairs(v)
-        w[key] += diagME(h, key)*val
-        for (add,elem) in hops(h, key)
-            w[add] += elem*val
-        end
-    end
-    return w
-end
-
-# three argument version
-function LinearAlgebra.mul!(w::AbstractDVec, h::AbstractHamiltonian, v::AbstractDVec)
-    empty!(w)
-    for (key,val) in pairs(v)
-        w[key] += diagME(h, key)*val
-        for (add,elem) in hops(h, key)
-            w[add] += elem*val
-        end
-    end
-    return w
-end
 
 (h::AbstractHamiltonian)(v) = h * v
 (h::AbstractHamiltonian)(w, v) = mul!(w, h, v)
@@ -67,6 +55,17 @@ BitStringAddresses.num_modes(h::AbstractHamiltonian) = num_modes(starting_addres
 
 Compute the diagonal matrix element of the linear operator `ham` at
 address `add`.
+
+# Example
+
+```jldoctest
+julia> addr = BoseFS((3, 2, 1));
+
+julia> H = HubbardMom1D(addr);
+
+julia> diagME(H, addr)
+8.666666666666664
+```
 """
 diagME
 
@@ -74,6 +73,17 @@ diagME
     numOfHops(ham, add)
 
 Compute the number of number of reachable configurations from address `add`.
+
+# Example
+
+```jldoctest
+julia> addr = BoseFS((3, 2, 1));
+
+julia> H = HubbardMom1D(addr);
+
+julia> numOfHops(H, addr)
+10
+```
 """
 numOfHops
 
@@ -82,21 +92,45 @@ numOfHops
 
 Compute matrix element of `hamiltonian` and new address of a single hop from
 address `add` with integer index `chosen`.
+
+# Example
+
+```jldoctest
+julia> addr = BoseFS((3, 2, 1));
+
+julia> H = HubbardMom1D(addr);
+
+julia> hop(H, addr, 3)
+(BoseFS{6,3}((2, 1, 3)), 1.0)
+```
 """
 hop
 
 """
-Approximate formula for log of binomial coefficient. Source:
-  <https://en.wikipedia.org/wiki/Binomial_coefficient#Bounds_and_asymptotic_formulas>
+Approximate formula for log of binomial coefficient. [Source](https://en.wikipedia.org/wiki/Binomial_coefficient#Bounds_and_asymptotic_formulas)
 """
 logbinomialapprox(n,k) =
     (n+0.5)*log((n+0.5)/(n-k+0.5))+k*log((n-k+0.5)/k) - 0.5*log(2π*k)
 
 """
-    dimension(::Type{T<:Integer}, h)
+    dimension(::Type{T}, h)
 
 Return the dimension of Hilbert space as `T`. If the result does not fit into `T`, return
-`nothing`
+`nothing`. If `T<:AbstractFloat`, an approximate value computed with the improved
+Stirling formula may be returned instead.
+
+# Examples
+
+```jldoctest
+julia> dimension(HubbardMom1D(BoseFS((1,2,3))))
+28
+julia> dimension(HubbardMom1D(nearUniform(BoseFS{200,100})))
+
+julia> dimension(Float64, HubbardMom1D(nearUniform(BoseFS{200,100})))
+1.3862737677578232e81
+julia> dimension(BigInt, HubbardMom1D(nearUniform(BoseFS{200,100})))
+1386083821086188248261127842108801860093488668581216236221011219101585442774669540
+```
 """
 function dimension(::Type{T}, ::BoseFS{N,M}) where {N,M,T<:Integer}
     try
@@ -105,12 +139,6 @@ function dimension(::Type{T}, ::BoseFS{N,M}) where {N,M,T<:Integer}
         return nothing
     end
 end
-"""
-    dimension(::Type{T<:AbstractFloat}, h)
-
-Return the dimension of Hilbert space as `T`. The exact result is returned if the value is
-smaller than 2^53. Otherwise, an improved Stirling formula is used.
-"""
 function dimension(::Type{T}, ::BoseFS{N,M}) where {N,M,T<:AbstractFloat}
     n = N + M - 1
     k = N
@@ -130,20 +158,33 @@ BitStringAddresses.nearUniform(h::AbstractHamiltonian) = nearUniform(typeof(star
     starting_address(h)
 
 Return a starting address for Hamiltonian `h`.
+
+# Example
+
+```jldoctest
+julia> addr = BoseFS((3, 2, 1));
+
+julia> H = HubbardMom1D(addr);
+
+julia> addr == starting_address(H)
+true
+```
 """
 starting_address
 
 """
     Hamiltonians.LOStructure(op::AbstractHamiltonian)
     Hamiltonians.LOStructure(typeof(op))
-`LOStructure` speficies properties of the linear operator `op`. If a special
-structure is known this can speed up calculations. Implemented structures are:
+
+`LOStructure` speficies properties of the linear operator `op`. If a special structure is
+known this can speed up calculations. Implemented structures are:
 
 * `Hamiltonians.HermitianLO` The operator is complex and hermitian or real and symmetric.
 * `Hamiltonians.ComplexLO` The operator has no known specific structure.
 
-In order to define this trait for a new linear operator type, define a method
-for `LOStructure(::Type{T}) where T <: MyNewLOType = …`.
+In order to define this trait for a new linear operator type, define a method for
+`LOStructure(::Type{<:MyNewLOType}) = …`.
+
 """
 abstract type LOStructure end
 
@@ -154,165 +195,14 @@ struct ComplexLO <: LOStructure end
 LOStructure(op::AbstractHamiltonian) = LOStructure(typeof(op))
 LOStructure(::Type{<:AbstractHamiltonian}) = ComplexLO()
 
-LinearAlgebra.adjoint(op::AbstractHamiltonian) = h_adjoint(LOStructure(op), op)
-
 """
-    h_adjoint(los::LOStructure, op::AbstractHamiltonian)
-Represent the adjoint of an `AbstractHamiltonian`. Extend this method to define
-custom adjoints.
-"""
-function h_adjoint(los::LOStructure, op) # default
-    throw(ErrorException("`adjoint()` not defined for `AbstractHamiltonian`s with `LOStructure` `$(typeof(los))`. Is your Hamiltonian hermitian?"))
-    return op
-end
+    rayleigh_quotient(H, v)
 
-h_adjoint(::HermitianLO, op) = op # adjoint is known
-
-"""
-    dot(x, LO::AbstractHamiltonian, v)
-Evaluate `x⋅LO(v)` minimizing memory allocations.
-"""
-function LinearAlgebra.dot(x::AbstractDVec, LO::AbstractHamiltonian, v::AbstractDVec)
-    return dot_w_trait(LOStructure(LO), x, LO, v)
-end
-# specialised method for UniformProjector
-function LinearAlgebra.dot(::UniformProjector, LO::AbstractHamiltonian{T}, v::AbstractDVec{K,T2}) where {K, T, T2}
-    result = zero(promote_type(T,T2))
-    for (key,val) in pairs(v)
-        result += diagME(LO, key) * val
-        for (add,elem) in hops(LO, key)
-            result += elem * val
-        end
-    end
-    return result
-end
-
-"""
-    Hamiltonians.dot_w_trait(::LOStructure, x, LO::AbstractHamiltonian, v)
-Internal function for making use of the `AbstractHamiltonian` trait `LOStructure`.
-"""
-dot_w_trait(::LOStructure, x, LO::AbstractHamiltonian, v) = dot_from_right(x,LO,v)
-# default for LOs without special structure: keep order
-
-function dot_w_trait(::HermitianLO, x, LO::AbstractHamiltonian, v)
-    if length(x) < length(v)
-        return conj(dot_from_right(v,LO,x)) # turn args around to execute faster
-    else
-        return dot_from_right(x,LO,v) # original order
-    end
-end
-
-"""
-    Hamiltonians.dot_from_right(x, LO, v)
-Internal function evaluates the 3-argument `dot()` function in order from right
-to left.
-"""
-function dot_from_right(x::AbstractDVec{K,T1}, LO::AbstractHamiltonian{T}, v::AbstractDVec{K,T2}) where {K, T,T1, T2}
-    # function LinearAlgebra.dot(x::AbstractDVec{K,T1}, LO::AbstractHamiltonian{T}, v::AbstractDVec{K,T2}) where {K, T,T1, T2}
-    result = zero(promote_type(T1,promote_type(T,T2)))
-    for (key,val) in pairs(v)
-        result += conj(x[key]) * diagME(LO, key) * val
-        for (add,elem) in hops(LO, key)
-            result += conj(x[add]) * elem * val
-        end
-    end
-    return result
-end
-
-
-"""
-    generateRandHop(ham, add)
-    generateRandHop(hops::Hops)
-
-Generate a single random excitation, i.e. choose from one of the accessible
-off-diagonal elements in the column corresponding to address `add` of
-the Hamiltonian matrix represented by `ham`. Alternatively, pass as
-argument an iterator over the accessible matrix elements.
-"""
-function generateRandHop(ham::AbstractHamiltonian, add)
-    # generic implementation of a random excitation generator drawing from
-    # a uniform distribution
-    # draws a random linked site and returns the site's address and generation
-    # probability
-    nl = numOfHops(ham, add) # check how many sites we could hop to
-    chosen = cRand(1:nl) # choose one of them
-    #chosen = _nearlydivisionless(nl) + 1 # choose one of them
-    # using a faster random number algorithm
-    naddress, melem = hop(ham, add, chosen)
-    return naddress, 1 ./nl, melem
-    # return new address, generation probability, and matrix element
-end
-
-function Base.getindex(ham::AbstractHamiltonian{T}, address1, address2) where T
-    # calculate the matrix element when only two bitstring addresses are given
-    # this is NOT used for the QMC algorithm and is currenlty not used either
-    # for building the matrix for conventional diagonalisation.
-    # Only used for verifying matrix.
-    # This will be slow and inefficient. Avoid using for larger Hamiltonians!
-    address1 == address2 && return diagME(ham, address1) # diagonal
-    for (add,val) in hops(ham, address2) # off-diag column as iterator
-        add == address1 && return val # found address1
-    end
-    return zero(T) # address1 not found
-end # getindex(ham)
-
-# A handy function that helps make use of the `AbstractHamiltonian` technology.
-"""
-    rayleigh_quotient(lo, v)
-Compute
 ```math
-\\frac{⟨ v | lo | v ⟩}{⟨ v|v ⟩}
+\\frac{⟨ v | H | v ⟩}{⟨ v|v ⟩}
 ```
 """
 rayleigh_quotient(lo, v) = dot(v, lo, v)/norm(v)^2
-
-
-"""
-    sm, basis = build_sparse_matrix_from_LO(ham::AbstractHamiltonian, add; nnzs = 0)
-Create a sparse matrix `sm` of all reachable matrix elements of a linear operator `ham`
-starting from the address `add`. The vector `basis` contains the addresses of basis configurations.
-Providing the number `nnzs` of expected calculated matrix elements may improve performance.
-"""
-function build_sparse_matrix_from_LO(ham::AbstractHamiltonian, fs; nnzs = 0)
-    adds = [fs] # list of addresses of length linear dimension of matrix
-    I = Vector{Int}(undef,0) # row indices, length nnz
-    J = Vector{Int}(undef,0) # column indices, length nnz
-    V = Vector{eltype(ham)}(undef,0) # values, length nnz
-    if nnzs > 0
-        sizehint!(I, nnzs)
-        sizehint!(J, nnzs)
-        sizehint!(V, nnzs)
-    end
-
-    k = 0 # 1:nnz, in principle, but possibly more as several contributions to a matrix element may occur
-    i = 0 # 1:dim, column of matrix
-    while true # loop over columns of the matrix
-        k += 1
-        i += 1 # next column
-        i > length(adds) && break
-        add = adds[i] # new address from list
-        # compute and push diagonal matrix element
-        melem = diagME(ham, add)
-        push!(I, i)
-        push!(J, i)
-        push!(V, melem)
-        for (nadd, melem) in hops(ham, add) # loop over rows
-            k += 1
-            j = findnext(a->a == nadd, adds, 1) # find index of `nadd` in `adds`
-            if isnothing(j)
-                # new address: increase dimension of matrix by adding a row
-                push!(adds, nadd)
-                j = length(adds) # row index points to the new element in `adds`
-            end
-            # new nonzero matrix element
-            push!(I, i)
-            push!(J, j)
-            push!(V, melem)
-        end
-    end
-    return sparse(I,J,V), adds
-    # when the index pair `(i,j)` occurs mutiple times in `I` and `J` the elements are added.
-end
 
 """
     TwoComponentHamiltonian{T} <: AbstractHamiltonian{T}
@@ -323,19 +213,42 @@ two different species. At least the following fields should be present:
 * `ha` Hamiltonian for species A
 * `hb` Hamiltonian for species B
 
-Methods that need to be implemented:
-* [`numOfHops(lo::AbstractHamiltonian, address)`](@ref) - number of off-diagonal matrix elements
-* [`hop(lo::AbstractHamiltonian, address, chosen::Integer)`](@ref) - access an off-diagonal m.e. by index `chosen`
-* [`diagME(lo::AbstractHamiltonian, address)`](@ref) - diagonal matrix element
-Optional:
-* [`Hamiltonians.LOStructure(::Type{typeof(lo)})`](@ref) - can speed up deterministic calculations if `HermitianLO`
+See [`AbstractHamiltonian`](@ref) for a list of methods that need to be defined.
 
-Provides:
-* [`dimension(::Type{T}, ::AbstractHamiltonian)`](@ref)
-
+Provides and implementation of [`dimension`](@ref).
 """
 abstract type TwoComponentHamiltonian{T} <: AbstractHamiltonian{T} end
 
 function dimension(::Type{T}, h::TwoComponentHamiltonian) where {T}
     return dimension(T, h.ha) * dimension(T, h.hb)
 end
+
+"""
+    momentum(ham::AbstractHamiltonian)
+
+Momentum as a linear operator in Fock space. Pass a Hamiltonian `ham` in order to convey
+information about the Fock basis.
+
+Note: `momentum` is currently only defined on [`HubbardMom1D`](@ref).
+
+# Example
+
+```jldoctest
+julia> using Rimu.Hamiltonians: MomentumMom1D # MomentumMom1D is unexported
+
+julia> add = BoseFS((1, 0, 2, 1, 2, 1, 1, 3));
+
+julia> ham = HubbardMom1D(add; u = 2.0, t = 1.0);
+
+julia> mom = momentum(ham);
+
+julia> diagME(mom, add) # calculate the momentum of a single configuration
+-1.5707963267948966
+
+julia> v = DVec(add => 10; capacity=1000);
+
+julia> rayleigh_quotient(mom, v) # momentum expectation value for state vector `v`
+-1.5707963267948966
+```
+"""
+momentum
